+++ ledger/participant-state/kvutils/tools/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/v1/Main.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.tools.driver
+package com.daml.ledger.participant.state.kvutils.tools.integritycheck.v1
 
 import java.io.DataInputStream
 import java.nio.file.{Files, Paths}
@@ -19,25 +19,7 @@ import com.daml.metrics.{JvmMetricSet, Metrics}
 import scala.collection.JavaConverters._
 import scala.util.Try
 
-final class Timer {
-  private var last = 0L
-  private var total = 0L
-
-  def time[T](act: => T): T = {
-    val start = System.nanoTime()
-    val result: T = act // force the thunk
-    val end = System.nanoTime()
-    last = end - start
-    total += last
-    result
-  }
-
-  def lastMs: Long = TimeUnit.NANOSECONDS.toMillis(last)
-
-  def totalMs: Long = TimeUnit.NANOSECONDS.toMillis(total)
-}
-
-object IntegrityCheck extends App {
+object Main extends App {
   if (args.length != 1) {
     println("usage: integrity-check <ledger dump file>")
     println(
