+++ ledger/participant-state/kvutils/tools/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/v2/IntegrityChecker.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.tools.export
+package com.daml.ledger.participant.state.kvutils.tools.integritycheck.v2
 
 import java.util.concurrent.TimeUnit
 
@@ -16,7 +16,7 @@ import com.daml.ledger.participant.state.kvutils.export.{
   NoOpLedgerDataExporter,
   WriteSet
 }
-import com.daml.ledger.participant.state.kvutils.tools._
+import com.daml.ledger.participant.state.kvutils.tools.integritycheck.v2.Color.color
 import com.daml.ledger.validator.LedgerStateOperations.{Key, Value}
 import com.daml.ledger.validator.batch.{
   BatchedSubmissionValidator,
