+++ ledger/participant-state/kvutils/tools/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/v2/WriteRecordingLedgerStateOperations.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.tools.export
+package com.daml.ledger.participant.state.kvutils.tools.integritycheck.v2
 
 import com.daml.ledger.participant.state.kvutils.export.WriteSet
 import com.daml.ledger.validator.LedgerStateOperations
